Extract `sinon.mock` into a CommonJS module.
Extract, `sinon.expectation`, an inner module into its own CommonJS module (`mock-expectation`).  This is still exported on the `sinon` object to satisfy tests.
