+++ lib/sinon/mock.js
@@ -8,18 +8,14 @@
  */
 "use strict";
 
-var sinon = require("./util/core");
+var mockExpectation = require("./mock-expectation");
 var spyCallToString = require("./call").toString;
-var spyInvoke = require("./spy").invoke;
 var extend = require("./extend");
-var assert = require("./assert");
-var stub = require("./stub");
 var match = require("./match");
 var deepEqual = require("./util/core/deep-equal").use(match);
-var timesInWords = require("./util/core/times-in-words");
 var wrapMethod = require("./util/core/wrap-method");
 
-var push = [].push;
+var push = Array.prototype.push;
 
 function mock(object) {
     // if (typeof console !== undefined && console.warn) {
@@ -27,7 +23,7 @@ function mock(object) {
     // }
 
     if (!object) {
-        return sinon.expectation.create("Anonymous mock");
+        return mockExpectation.create("Anonymous mock");
     }
 
     return mock.create(object);
@@ -90,7 +86,7 @@ extend(mock, {
             push.call(this.proxies, method);
         }
 
-        var expectation = sinon.expectation.create(method);
+        var expectation = mockExpectation.create(method);
         push.call(this.expectations[method], expectation);
 
         return expectation;
@@ -124,9 +120,9 @@ extend(mock, {
         this.restore();
 
         if (messages.length > 0) {
-            sinon.expectation.fail(messages.concat(met).join("\n"));
+            mockExpectation.fail(messages.concat(met).join("\n"));
         } else if (met.length > 0) {
-            sinon.expectation.pass(messages.concat(met).join("\n"));
+            mockExpectation.pass(messages.concat(met).join("\n"));
         }
 
         return true;
@@ -176,278 +172,8 @@ extend(mock, {
             args: args
         }));
 
-        sinon.expectation.fail(messages.join("\n"));
+        mockExpectation.fail(messages.join("\n"));
     }
 });
 
-var slice = Array.prototype.slice;
-
-function callCountInWords(callCount) {
-    if (callCount === 0) {
-        return "never called";
-    }
-
-    return "called " + timesInWords(callCount);
-}
-
-function expectedCallCountInWords(expectation) {
-    var min = expectation.minCalls;
-    var max = expectation.maxCalls;
-
-    if (typeof min === "number" && typeof max === "number") {
-        var str = timesInWords(min);
-
-        if (min !== max) {
-            str = "at least " + str + " and at most " + timesInWords(max);
-        }
-
-        return str;
-    }
-
-    if (typeof min === "number") {
-        return "at least " + timesInWords(min);
-    }
-
-    return "at most " + timesInWords(max);
-}
-
-function receivedMinCalls(expectation) {
-    var hasMinLimit = typeof expectation.minCalls === "number";
-    return !hasMinLimit || expectation.callCount >= expectation.minCalls;
-}
-
-function receivedMaxCalls(expectation) {
-    if (typeof expectation.maxCalls !== "number") {
-        return false;
-    }
-
-    return expectation.callCount === expectation.maxCalls;
-}
-
-function verifyMatcher(possibleMatcher, arg) {
-    var isMatcher = match && match.isMatcher(possibleMatcher);
-
-    return isMatcher && possibleMatcher.test(arg) || true;
-}
-
-sinon.expectation = {
-    minCalls: 1,
-    maxCalls: 1,
-
-    create: function create(methodName) {
-        var expectation = extend(stub.create(), sinon.expectation);
-        delete expectation.create;
-        expectation.method = methodName;
-
-        return expectation;
-    },
-
-    invoke: function invoke(func, thisValue, args) {
-        this.verifyCallAllowed(thisValue, args);
-
-        return spyInvoke.apply(this, arguments);
-    },
-
-    atLeast: function atLeast(num) {
-        if (typeof num !== "number") {
-            throw new TypeError("'" + num + "' is not number");
-        }
-
-        if (!this.limitsSet) {
-            this.maxCalls = null;
-            this.limitsSet = true;
-        }
-
-        this.minCalls = num;
-
-        return this;
-    },
-
-    atMost: function atMost(num) {
-        if (typeof num !== "number") {
-            throw new TypeError("'" + num + "' is not number");
-        }
-
-        if (!this.limitsSet) {
-            this.minCalls = null;
-            this.limitsSet = true;
-        }
-
-        this.maxCalls = num;
-
-        return this;
-    },
-
-    never: function never() {
-        return this.exactly(0);
-    },
-
-    once: function once() {
-        return this.exactly(1);
-    },
-
-    twice: function twice() {
-        return this.exactly(2);
-    },
-
-    thrice: function thrice() {
-        return this.exactly(3);
-    },
-
-    exactly: function exactly(num) {
-        if (typeof num !== "number") {
-            throw new TypeError("'" + num + "' is not a number");
-        }
-
-        this.atLeast(num);
-        return this.atMost(num);
-    },
-
-    met: function met() {
-        return !this.failed && receivedMinCalls(this);
-    },
-
-    verifyCallAllowed: function verifyCallAllowed(thisValue, args) {
-        if (receivedMaxCalls(this)) {
-            this.failed = true;
-            sinon.expectation.fail(this.method + " already called " + timesInWords(this.maxCalls));
-        }
-
-        if ("expectedThis" in this && this.expectedThis !== thisValue) {
-            sinon.expectation.fail(this.method + " called with " + thisValue + " as thisValue, expected " +
-                this.expectedThis);
-        }
-
-        if (!("expectedArguments" in this)) {
-            return;
-        }
-
-        if (!args) {
-            sinon.expectation.fail(this.method + " received no arguments, expected " +
-                sinon.format(this.expectedArguments));
-        }
-
-        if (args.length < this.expectedArguments.length) {
-            sinon.expectation.fail(this.method + " received too few arguments (" + sinon.format(args) +
-                "), expected " + sinon.format(this.expectedArguments));
-        }
-
-        if (this.expectsExactArgCount &&
-            args.length !== this.expectedArguments.length) {
-            sinon.expectation.fail(this.method + " received too many arguments (" + sinon.format(args) +
-                "), expected " + sinon.format(this.expectedArguments));
-        }
-
-        for (var i = 0, l = this.expectedArguments.length; i < l; i += 1) {
-
-            if (!verifyMatcher(this.expectedArguments[i], args[i])) {
-                sinon.expectation.fail(this.method + " received wrong arguments " + sinon.format(args) +
-                    ", didn't match " + this.expectedArguments.toString());
-            }
-
-            if (!deepEqual(this.expectedArguments[i], args[i])) {
-                sinon.expectation.fail(this.method + " received wrong arguments " + sinon.format(args) +
-                    ", expected " + sinon.format(this.expectedArguments));
-            }
-        }
-    },
-
-    allowsCall: function allowsCall(thisValue, args) {
-        if (this.met() && receivedMaxCalls(this)) {
-            return false;
-        }
-
-        if ("expectedThis" in this && this.expectedThis !== thisValue) {
-            return false;
-        }
-
-        if (!("expectedArguments" in this)) {
-            return true;
-        }
-
-        args = args || [];
-
-        if (args.length < this.expectedArguments.length) {
-            return false;
-        }
-
-        if (this.expectsExactArgCount &&
-            args.length !== this.expectedArguments.length) {
-            return false;
-        }
-
-        for (var i = 0, l = this.expectedArguments.length; i < l; i += 1) {
-            if (!verifyMatcher(this.expectedArguments[i], args[i])) {
-                return false;
-            }
-
-            if (!deepEqual(this.expectedArguments[i], args[i])) {
-                return false;
-            }
-        }
-
-        return true;
-    },
-
-    withArgs: function withArgs() {
-        this.expectedArguments = slice.call(arguments);
-        return this;
-    },
-
-    withExactArgs: function withExactArgs() {
-        this.withArgs.apply(this, arguments);
-        this.expectsExactArgCount = true;
-        return this;
-    },
-
-    on: function on(thisValue) {
-        this.expectedThis = thisValue;
-        return this;
-    },
-
-    toString: function () {
-        var args = (this.expectedArguments || []).slice();
-
-        if (!this.expectsExactArgCount) {
-            push.call(args, "[...]");
-        }
-
-        var callStr = spyCallToString.call({
-            proxy: this.method || "anonymous mock expectation",
-            args: args
-        });
-
-        var message = callStr.replace(", [...", "[, ...") + " " +
-            expectedCallCountInWords(this);
-
-        if (this.met()) {
-            return "Expectation met: " + message;
-        }
-
-        return "Expected " + message + " (" +
-            callCountInWords(this.callCount) + ")";
-    },
-
-    verify: function verify() {
-        if (!this.met()) {
-            sinon.expectation.fail(this.toString());
-        } else {
-            sinon.expectation.pass(this.toString());
-        }
-
-        return true;
-    },
-
-    pass: function pass(message) {
-        assert.pass(message);
-    },
-
-    fail: function fail(message) {
-        var exception = new Error(message);
-        exception.name = "ExpectationError";
-
-        throw exception;
-    }
-};
-
-sinon.mock = mock;
+module.exports = mock;
